Fix resize width
