+++ src/utils/index.ts
@@ -1,13 +1,10 @@
-import type { CSSProperties } from 'react';
-import clsx from 'clsx';
-
 import type { CalculatedColumn } from '../types';
-import { cellClassname, cellFrozenClassname, cellFrozenLastClassname } from '../style';
 
 export * from './colSpanUtils';
 export * from './domUtils';
 export * from './keyboardUtils';
 export * from './selectedCellUtils';
+export * from './styleUtils';
 
 export const { min, max, round, floor, sign, abs } = Math;
 
@@ -19,37 +16,15 @@ export function assertIsValidKeyGetter<R, K extends React.Key>(
   }
 }
 
-export function getRowStyle(rowIdx: number, height?: number): CSSProperties {
-  if (height !== undefined) {
-    return {
-      '--rdg-grid-row-start': rowIdx,
-      '--rdg-row-height': `${height}px`
-    } as unknown as CSSProperties;
+export function clampColumnWidth<R, SR>(
+  width: number,
+  { minWidth, maxWidth }: CalculatedColumn<R, SR>
+): number {
+  width = max(width, minWidth);
+
+  if (typeof maxWidth === 'number' && maxWidth > minWidth) {
+    return min(width, maxWidth);
   }
-  return { '--rdg-grid-row-start': rowIdx } as unknown as CSSProperties;
-}
 
-export function getCellStyle<R, SR>(
-  column: CalculatedColumn<R, SR>,
-  colSpan?: number
-): React.CSSProperties {
-  return {
-    gridColumnStart: column.idx + 1,
-    gridColumnEnd: colSpan !== undefined ? `span ${colSpan}` : undefined,
-    insetInlineStart: column.frozen ? `var(--rdg-frozen-left-${column.idx})` : undefined
-  };
-}
-
-export function getCellClassname<R, SR>(
-  column: CalculatedColumn<R, SR>,
-  ...extraClasses: Parameters<typeof clsx>
-): string {
-  return clsx(
-    cellClassname,
-    {
-      [cellFrozenClassname]: column.frozen,
-      [cellFrozenLastClassname]: column.isLastFrozenColumn
-    },
-    ...extraClasses
-  );
+  return width;
 }
